// openmetadata-ui/src/main/resources/ui/playwright/support/interfaces/ServiceForm.interface.ts
/*
 *  Copyright 2025 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */
import { Page } from '@playwright/test';

export interface SupersetFormType {
  hostPort: string;
  connectionType: string;
  connection: {
    username: string;
    password: string;
    provider?: string;
    hostPort?: string;
    database?: string;
    scheme?: string;
  };
}

export interface FillSupersetFormProps extends SupersetFormType {
  page: Page;
}